lan and user: Don't leak a session when changing own password
When a user's password is changed in a session that belongs to
that user, some devices immediately switch to authenticating using
the new password within the session, while others only use the
new password for future sessions.  This causes the "Close Session"
command to fail for some devices when changing one's own password
with ipmitool, resulting in the session remaining open until it
times out.

This patch causes ipmitool first to attempt to close the session using
the new password, then, if that fails, to try using the password
that was originally used to open the session.