SW-2015: extended health_monitor status command to return a string.
Example: when RMCP, ICMP, and route_check monitors are on, the
status string will look like:

	on: rmcp icmp route_check

For config_ping the moniker is 'cfgp', and for mcam_check it's
'mcam_check'.

When health_monitoring is off, (e.g. at the outset, or
after 'cxoem fabric config health_monitor stop')
the status string returned will be:

	off