SW-2377: Lanplus session privilege should be set no higher than user max
By default, lanplus session are open with ADMINISTRATOR privileges
requested.  Later, once the session is open, an explicit attempt is made to
set the privilege level.  If the requested level (again, default is ADMIN)
is higher than the user's maximum level, the request is denied and
ipmitool throws out the session.

The correct thing to do is determine the minimum of the user's maximum
privilege level and the current requested prvileged level and request
elevation of session privileges to _that_ level.